added controllers and routes
