+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "mocha ./test/test.js",
     "dev": "nodemon --exec babel-node --presets babel-preset-env ./server.js"
   },
   "repository": {
