+++ public/index.html
@@ -5,8 +5,14 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <title>Politico</title>
+    <style>
+        span {
+            color: blue;
+        }
+    </style>
 </head>
 <body>
-    <h1>Server for Politico is now active</h1>
+    <h1>Server for Politico is running</h1>
+    <p>To get parties, or offices - navigate to <span>ap1/v1/</span> and add: <em>parties</em> or <em>offices</em></p>
 </body>
 </html>
